+++ packages/lottie/src/index.ts
@@ -2,3 +2,4 @@ import { LottieView } from './GraniteLottieView';
 export { LottieView };
 export default LottieView;
 export * from './types';
+export type { OnAnimationFailureEvent, OnAnimationFinishEvent } from './GraniteLottieViewNativeComponent';