+++ packages/react-native/CHANGELOG.md
@@ -1,11 +1,16 @@
 # @granite-js/react-native
 
+<<<<<<< HEAD
 ## 1.0.0-next.87
+=======
+## 0.1.34-next.46
+>>>>>>> feature/rn81
 
 ### Patch Changes
 
 - fix
 - Updated dependencies
+<<<<<<< HEAD
   - @granite-js/plugin-core@1.0.0-next.87
   - @granite-js/style-utils@1.0.0-next.87
   - @granite-js/blur-view@1.0.0-next.87
@@ -18,11 +23,26 @@
   - @granite-js/cli@1.0.0-next.87
 
 ## 1.0.0-next.86
+=======
+  - @granite-js/plugin-core@0.1.34-next.46
+  - @granite-js/style-utils@0.1.34-next.46
+  - @granite-js/blur-view@0.1.34-next.46
+  - @granite-js/lottie@0.1.34-next.46
+  - @granite-js/native@0.1.34-next.46
+  - @granite-js/image@0.1.34-next.46
+  - @granite-js/mpack@0.1.34-next.46
+  - @granite-js/video@0.1.34-next.46
+  - @granite-js/jest@0.1.34-next.46
+  - @granite-js/cli@0.1.34-next.46
+
+## 0.1.34-next.45
+>>>>>>> feature/rn81
 
 ### Patch Changes
 
 - fix
 - Updated dependencies
+<<<<<<< HEAD
   - @granite-js/plugin-core@1.0.0-next.86
   - @granite-js/style-utils@1.0.0-next.86
   - @granite-js/blur-view@1.0.0-next.86
@@ -50,6 +70,18 @@
   - @granite-js/video@1.0.0-next.85
   - @granite-js/jest@1.0.0-next.85
   - @granite-js/cli@1.0.0-next.85
+=======
+  - @granite-js/plugin-core@0.1.34-next.45
+  - @granite-js/style-utils@0.1.34-next.45
+  - @granite-js/blur-view@0.1.34-next.45
+  - @granite-js/lottie@0.1.34-next.45
+  - @granite-js/native@0.1.34-next.45
+  - @granite-js/image@0.1.34-next.45
+  - @granite-js/mpack@0.1.34-next.45
+  - @granite-js/video@0.1.34-next.45
+  - @granite-js/jest@0.1.34-next.45
+  - @granite-js/cli@0.1.34-next.45
+>>>>>>> feature/rn81
 
 ## 0.1.34-next.44